Implement site id in IntVector, GCounter and PNCounter
Previously, the user could specify the 'position' to update in the
vector. Now, each call to `make ()` will generate a random site id, and
subsequent calls to update / incr / decr will only update that specific
position of the vector. Merge operations do not modify the id.

Note that by default, `make ()`-generated ids are modulo 10, so there is
a high probability of collision. To preven this, use the `make_in_range`
constructor, supplying the appropiate modulo value.

If two nodes have different state size, the smaller one will grow to the
larger size on merge, padding the state with zeros.